requirements.txt
fastapi
uvicorn[standard]
rembg
pillow
onnxruntime
python-multipart

